backend/pyproject.toml
[project]
name = "closest-text-word-embeddings"
version = "0.1.0"
description = "Exploration in word embeddings"
authors = [
    {name = "Dhruv Sharma",email = "dhruv.sharma@polytechnique.org"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
   "faiss-cpu (>=1.11.0,<2.0.0)",
   "scikit-learn (>=1.6.1,<2.0.0)",
   "sentence-transformers (>=4.1.0,<5.0.0)",
   "fastapi (>=0.115.12,<0.116.0)",
   "typeguard (>=4.4.2,<5.0.0)",
   "mypy (>=1.15.0,<2.0.0)",
   "pydantic-settings (>=2.0.0,<3.0.0)",
   "loguru (>=0.7.0,<0.8.0)",
   "pytest (>=8.3.5,<9.0.0)",
    "numpy (>=2.2.6,<3.0.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "datasets (>=3.6.0,<4.0.0)",
    "nltk (>=3.9.1,<4.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest-asyncio (>=0.26.0,<0.27.0)"
]

[tool.poetry]
packages = [
  { include = "api",          from = "src" },
  { include = "core",         from = "src" },
  { include = "data",         from = "src" },
  { include = "experiments",  from = "src" },
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

